apps/agent/pyproject.toml
[project]
name = "code-gen-agent"
version = "0.0.1"
requires-python = ">=3.12.0,<3.14"
dependencies = [
    "spoon-ai-sdk>=0.3.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0"
]

